+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - New #29: Add `sendVideoNote` method.
 - New #31: Add `sendMediaGroup` method and `InputMediaAnimation`, `InputMediaAudio`, `InputMediaDocument`,
   `InputMediaPhoto`, `InputMediaVideo` types.
+- New #32: Add `setMessageReaction` method and `ReactionType::toRequestArray()` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.